Refactor: `shuffle` 유틸 분리
