+++ src/js/utils/index.js
@@ -1,0 +1,2 @@
+export * from './shuffle.js';
+export * from './splitToNumberArray.js';
