+++ src/js/utils/shuffle.js
@@ -1,0 +1,1 @@
+export const shuffle = (arr) => arr.slice().sort(() => Math.random() - 0.5);
