+++ src/js/utils/splitToNumberArray.js
@@ -1,7 +1,5 @@
-const splitToNumberArray = (input, separator = ',') =>
+export const splitToNumberArray = (input, separator = ',') =>
   input
     .split(separator)
     .map((number) => number.trim())
     .map(Number);
-
-export default splitToNumberArray;
